% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{applyKMeans}
\alias{applyKMeans}
\title{Applies K-Means algorithm to the given data data.mat}
\usage{
applyKMeans(dataMatrix, K, maxIter)
}
\arguments{
\item{dataMatrix}{Dataset of type matrix (List of double) and shape (n x d), where n is the number of observations and d is the dimension of the dataset.}

\item{K}{The number of clusters.}

\item{maxIter}{Maximum number of iterations of Kmeans clustering algorithm.}
}
\value{
The centroids of the clusters.

  For description of how KMeans clustering algorithm works, please see the following url.
  \url{https://en.wikipedia.org/wiki/K-means_clustering}
}
\description{
\code{applyKMeans} returns the the cluster centroids.
}
\examples{
applyKMeans(as.matrix(iris[,-5]), 3, 300)
}
